SilverStripe 5 compatibility (#16)
* 🐛  (data mapping) uses up-to-date methods where available. Removes others

This is step one of making this module compatible with Embed 4.x which is required by Silverstripe Framework 4.11. Fields which are no longer available by default are now not retrieved. It may be possible to retrieve this data still with
additional changes.

* ⬆️  (composer) makes embed dependency explicit and locked to version 4

Without an explicit dependency defined dependency this module is subject to fail when paired with a mismatched version of Silverstripe

* 🎨  (whitespace) changes spaces to tabs

* ⬆️  (dependencies) locks package to Silverstripe > 4.11

* Fixes thumbnail URL getting other attributes to work

* Fixes getting other attributes to work

* ⬆️  (dependencies) allows install in Silverstripe 5

* 🐛  (data-conversion) replaces deprecated method

* 🔥  (code) removes commented out lines

---------

Co-authored-by: sinanghareb <sinang@evanshunt.com>